+++ sources/matrix/inverting_matrices/cofactor.c
@@ -1,8 +1,8 @@
 #include <minirt.h>
 
-float	cofactor(t_matrix matrix, int i, int j)
+double	cofactor(t_matrix matrix, int i, int j)
 {
-	float	minor_result;
+	double	minor_result;
 
 	minor_result = minor(matrix, i, j);
 	if ((i + j) % 2 != 0)
